python/deterministic/requirements.txt
opentelemetry-api==0.13b0
opentelemetry-sdk==0.13b0
